+++ relay-cogs/src/cogs.rs
@@ -252,7 +252,7 @@ mod tests {
     use std::{collections::HashMap, time::Duration};
 
     use super::*;
-    use crate::TestRecorder;
+    use crate::test::TestRecorder;
 
     #[test]
     fn test_cogs_simple() {
