+++ relay-config/src/upstream.rs
@@ -158,7 +158,7 @@ impl FromStr for UpstreamDescriptor<'static> {
 relay_common::impl_str_serde!(UpstreamDescriptor<'static>, "a sentry upstream URL");
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::*;
 
     #[test]
